requirements.txt
Django==4.1.7
Pillow==9.4.0
asgiref==3.6.0
django-phonenumber-field==7.0.2
djangorestframework==3.14.0
env==0.1.0
envparse==0.2.0
phonenumbers==8.13.7
pip==23.0.1
psycopg2==2.9.5
pytz==2022.7.1
setuptools==67.6.0
sqlparse==0.4.3
tzdata==2022.7
wheel==0.38.4